상품 주문

+++ moda/src/main/java/com/moda/moda/order/OrderVO.java
@@ -25,7 +25,7 @@ public class OrderVO {
     private int prodCode;//
 
     //주문 상세
-    private Long ordDtlNo;
+    private int ordDtlNo;
     private String prodName;
     private int prodQty;
